Safari-611.1.10/LayoutTests/inspector/heap/tracking-expected.txt

Tests that Heap.startTracking and Heap.stopTracking trigger trackingStart and trackingComplete events with expected data.


== Running test suite: Heap.Tracking
-- Running test case: Heap.Tracking.StartAndStopTrackingIncludeSnapshots
Heap.trackingStart
PASS: Should have a timestamp when starting.
PASS: Should have snapshotData when starting.
Heap.trackingComplete
PASS: Should have a timestamp when stopping.
PASS: Should have snapshotData when stopping.
